Move lp fixture into conftest
"conftest" is another bit of pytest "magic"; everything in conftest
appears in the namespace of all test modules as if it were implicitly
imported (because, effectively, it is).

The only reason for doing this is the next commit will add tests for the
lp_cli module and the lp fixture is equally useful / applicable to that
module so it makes sense to make it universally available to all tests.